create app module[#465]

+++ app/src/main/kotlin/com/stone/fridge/data/pagination/RecipePagingSource.kt
@@ -38,7 +38,7 @@ class RecipePagingSource(
             }
 
             val response = api.getRecipe(
-                page = loadKey.toInt(),
+                page = loadKey,
                 size = state.config.pageSize
             )
